add eslint, start adding gh pages cause ...
I was developing the toaster as a separate module and I was
npm linking in a webpack starter kit, but because of that
two instances of React were messing things up. For more details
do not go here: https://github.com/facebook/react/issues/13991

So I didn't want to bother, and I was going to make a gh-pages
anyways, so I'm going to test the component and other stuff
here and npmignore the stuff that I don't want.